+++ tests/testthat/test-55-modify-table1.R
@@ -7,10 +7,14 @@ test_that('Table1 is modified', {
     charset("ANSI") %>%
     creation_date("2019-01-01 10:00") %>%
     matrix("BEXSTA") %>%
-    decimals("1")
+    decimals("1") %>%
+    last_updated("2020-01-01 10:00") %>%
+    next_update(format("2022-01-01 10:00", format='%Y%m%d %H:%M'))
 
   expect_equal(charset(x), "ANSI")
   expect_equal(creation_date(x), "2019-01-01 10:00")
   expect_equal(matrix(x), "BEXSTA")
   expect_equal(decimals(x), "1")
+  expect_equal(last_updated(x), "2020-01-01 10:00")
+  expect_equal(next_update(x), "2022-01-01 10:00")
 })