+++ NEWS.md
@@ -3,6 +3,7 @@
 ## New features
 - Change all modifying functions to also be getters, so e.g. `stub(x)` returns
 current STUB variables.
+- Add new modifying functions: `last_updated()` and `next_update()`.
 
 ## Bug fixes and minor improvements
 - Validate px after using modifying functions.